+++ core-ui/src/main/java/com/agileburo/anytype/core_ui/widgets/text/TextInputWidget.kt
@@ -1,11 +1,14 @@
 package com.agileburo.anytype.core_ui.widgets.text
 
 import android.content.Context
+import android.text.TextWatcher
 import android.util.AttributeSet
 import androidx.appcompat.widget.AppCompatEditText
 import timber.log.Timber
 
 class TextInputWidget : AppCompatEditText {
+
+    private val watchers: MutableList<TextWatcher> = mutableListOf()
 
     var selectionDetector: ((IntRange) -> Unit)? = null
 
@@ -16,6 +19,21 @@ class TextInputWidget : AppCompatEditText {
         attrs,
         defStyle
     )
+
+    override fun addTextChangedListener(watcher: TextWatcher) {
+        watchers.add(watcher)
+        super.addTextChangedListener(watcher)
+    }
+
+    override fun removeTextChangedListener(watcher: TextWatcher) {
+        watchers.remove(watcher)
+        super.removeTextChangedListener(watcher)
+    }
+
+    fun clearTextWatchers() {
+        watchers.forEach { super.removeTextChangedListener(it) }
+        watchers.clear()
+    }
 
     override fun onSelectionChanged(selStart: Int, selEnd: Int) {
         Timber.d("New selection: $selStart - $selEnd")